# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OnlinePort",
        "version": "0.1"
    },
    "paths": {
...
        "/api/car/right-uses": {
            "get": {
                "tags": [
                    "Car Right Uses"
                ],
                "summary": "Get list of right use",
                "operationId": "bde32252d213017863983de8674d7579",
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "post": {
                "tags": [
                    "Car Right Uses"
                ],
                "summary": "Store right use",
                "operationId": "e1d2ecf17ce5573e05a917333fef2651",
                "parameters": [
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Собственность"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
        "/api/car/right-uses/{id}": {
            "put": {
                "tags": [
                    "Car Right Uses"
                ],
                "summary": "Update right use",
                "description": "Update right use",
                "operationId": "Update right use",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Собственность"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
        "/api/car/pass/right-uses/{id}": {
            "delete": {
                "tags": [
                    "Car Right Uses"
                ],
                "summary": "Summary",
                "description": "Delete ight use",
                "operationId": "Delete right use",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
...
    }
}